Implemented infrastructure for enabling and disabling buttons in Simulation Panel Other minor changes to code and GUI

+++ CompareSearchMethods/Model/SimulationResults.cs
@@ -4,7 +4,7 @@ namespace CompareSearchMethods.Model
 	public class SimulationResults : ISimulationResults
 	{
 		public int NoOfEntries { get; set; }
-		public int NoOfSearches { get; set; }
+		public long NoOfSearches { get; set; }
 		public double AvgNoOfIterations { get; set; }
 		public double AvgElapsedTime { get; set; }
 	}
